Add an ugly AI
Sketched out some terrible code for a basic minimax algorithm. Just
committing for posterity, as it *does* seem to mostly do the job
(finally), but definitely needs some refinement. There's an example
function for it that makes the tests fail, but I was basically just
using it to poke at the code, since I couldn't easily slap in a main()
function.

The most interesting part of the minimax ugliness is that it actually
uncovered a bug in func (Turn) Evaluate() (Board, bool). The top-level
for-loop wouldn't actually ever allow the first round of sowing if the
pit started out with 1 seed. Oops. This commit fixes that and adds a
regression test (found by running the Decision example!).